Update DiffPIR for imaginary images
