+++ datasets/fastmri.py
@@ -82,8 +82,6 @@ class Dataset(BaseDataset):
 
         x, y, z = test_dataset[0]
         dinv.utils.plot([x.unsqueeze(0), y.unsqueeze(0), z["mask"].unsqueeze(0)])
-        
-        breakpoint()
 
         return dict(
             train_dataset=train_dataset,
