+++ solvers/dpir.py
@@ -6,12 +6,11 @@ with safe_import_context() as import_ctx:
     import deepinv as dinv
     import numpy as np
     from deepinv.optim import BaseOptim
-    from deepinv.models import Denoiser
     from deepinv.optim.prior import PnP
     from deepinv.optim.data_fidelity import L2
     from deepinv.optim.optimizers import create_iterator
     from deepinv.optim.dpir import get_DPIR_params
-    from deepinv.models import DRUNet
+    from benchmark_utils.denoiser_2c import Denoiser_2c
 
 class Solver(BaseSolver):
     name = 'DPIR'
@@ -67,10 +66,11 @@ class Solver(BaseSolver):
     def get_result(self):
         return dict(model=self.model, model_name="DPIR", device=self.device)
 
+
 # Custom DPIR solver with 2 channels
 class DPIR_2C(BaseOptim):
     def __init__(self, sigma=0.1, device="cuda"):
-        prior = PnP(denoiser=DPIR_2C_Denoiser(in_channels=1, out_channels=1, pretrained="download", device=device))
+        prior = PnP(denoiser=Denoiser_2c(device=device))
         sigma_denoiser, stepsize, max_iter = get_DPIR_params(sigma)
         params_algo = {"stepsize": stepsize, "g_param": sigma_denoiser}
         super(DPIR_2C, self).__init__(
@@ -81,20 +81,3 @@ class DPIR_2C(BaseOptim):
             early_stop=False,
             params_algo=params_algo,
         )
-
-
-class DPIR_2C_Denoiser(Denoiser):
-    def __init__(self, *DRUNet_args, **DRUNet_kwargs):
-        super(DPIR_2C_Denoiser, self).__init__()
-        self.model_c1 = DRUNet(*DRUNet_args, **DRUNet_kwargs)
-        self.model_c2 = DRUNet(*DRUNet_args, **DRUNet_kwargs)
-    
-    def forward(self, y, sigma):
-        y1, y2 = torch.split(y, 1, dim=1)
-
-        x_hat_1 = self.model_c1(y1, sigma=sigma)
-        x_hat_2 = self.model_c2(y2, sigma=sigma)
-
-        x_hat = torch.cat([x_hat_1, x_hat_2], dim=1)
-        
-        return x_hat
